Update on "Factoring out the quantized mappings"
Differential Revision: [D16470907](https://our.internmc.facebook.com/intern/diff/D16470907)
