+++ torch/quantization/_mappings.py
@@ -4,24 +4,27 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 import torch
+import torch.nn as nn
+import torch.nn.qat as qat
+import torch.nn.quantized as nnq
 
-from .quantize import QuantStub
-from .quantize import DeQuantStub
+from ._wrappers import QuantStub
+from ._wrappers import DeQuantStub
 
 # Map for swapping float module to quantized ones.
 _DEFAULT_MODULE_MAPPING = {
-    torch.nn.Linear: torch.nn.quantized.Linear,
-    torch.nn.ReLU: torch.nn.quantized.ReLU,
-    torch.nn.Conv2d: torch.nn.quantized.Conv2d,
-    QuantStub: torch.nn.quantized.Quantize,
-    DeQuantStub: nntorch.nn.quantizedq.DeQuantize,
+    nn.Linear: nnq.Linear,
+    nn.ReLU: nnq.ReLU,
+    nn.Conv2d: nnq.Conv2d,
+    QuantStub: nnq.Quantize,
+    DeQuantStub: nnq.DeQuantize,
     # QAT modules:
-    torch.nn.qat.Linear: torch.nn.quantized.Linear,
-    torch.nn.qat.Conv2d: torch.nn.quantized.Conv2d,
+    qat.Linear: nnq.Linear,
+    qat.Conv2d: nnq.Conv2d,
 }
 
 # Map for swapping float module to qat modules.
 _DEFAULT_QAT_MODULE_MAPPING = {
-    torch.nn.Linear: torch.nn.qat.Linear,
-    torch.nn.Conv2d: torch.nn.qat.Conv2d,
+    nn.Linear: nn.qat.Linear,
+    nn.Conv2d: nn.qat.Conv2d,
 }
