+++ torch/quantization/quantize.py
@@ -1,11 +1,10 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 import torch.nn as nn
-import torch.nn.quantized as nnq
-import torch.nn.qat as qat
 import torch
 
 from ._mappings import _DEFAULT_MODULE_MAPPING
 from ._mappings import _DEFAULT_QAT_MODULE_MAPPING
+from ._wrappers import QuantWrapper
 
 def propagate_qconfig_helper(module, qconfig_dict, qconfig_parent=None, prefix=''):
     r"""This is a helper function for `propagate_qconfig`
@@ -82,29 +81,6 @@ def add_observer(module):
         module.add_module('observer', module.qconfig.activation())
         module.register_forward_hook(_observer_forward_hook)
 
-class QuantWrapper(nn.Module):
-    r"""A wrapper class that wraps the input module, adds QuantStub and
-    DeQuantStub and surround the call to module with call to quant and dequant
-    modules.
-
-    This is used by the `quantization` utility functions to add the quant and
-    dequant modules, before `convert` function `QuantStub` will just be observer,
-    it observes the input tensor, after `convert`, `QuantStub`
-    will be swapped to `nnq.Quantize` which does actual quantization. Similarly
-    for `DeQuantStub`.
-    """
-    def __init__(self, module):
-        super(QuantWrapper, self).__init__()
-        qconfig = module.qconfig if hasattr(module, 'qconfig') else None
-        self.add_module('quant', QuantStub(qconfig))
-        self.add_module('dequant', DeQuantStub(qconfig))
-        self.add_module('module', module)
-        self.train(module.training)
-
-    def forward(self, X):
-        X = self.quant(X)
-        X = self.module(X)
-        return self.dequant(X)
 
 def add_quant_dequant(module):
     r"""Wrap the leaf child module in QuantWrapper if it has a valid qconfig
@@ -153,33 +129,6 @@ def prepare_qat(model, qconfig_dict=None):
     model = convert(model, _DEFAULT_QAT_MODULE_MAPPING)
     return model
 
-class QuantStub(nn.Module):
-    r"""Quantize stub module, before calibration, this is same as an observer,
-    it will be swapped as `nnq.Quantize` in `convert`.
-
-    Args:
-        qconfig: quantization configuration for the tensor,
-            if qconfig is not provided, we will get qconfig from parent modules
-    """
-    def __init__(self, qconfig=None):
-        super(QuantStub, self).__init__()
-        if qconfig:
-            self.qconfig = qconfig
-
-    def forward(self, x):
-        return x
-
-class DeQuantStub(nn.Module):
-    r"""Dequantize stub module, before calibration, this is same as identity,
-    this will be swapped as `nnq.DeQuantize` in `convert`.
-    """
-    def __init__(self, qconfig=None):
-        super(DeQuantStub, self).__init__()
-        if qconfig:
-            self.qconfig = qconfig
-
-    def forward(self, x):
-        return x
 
 def quantize(model, run_fn, run_args, qconfig_dict=None):
     r"""Converts a float model to quantized model.
